Miguel Api
Subida dia 2 Miguel Api

+++ APICoches/src/main/java/com/example/demo/repository/PersonalRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Personal;
 
-public interface PersonalRepository extends JpaRepository<Personal, Long>{
-
+public interface PersonalRepository extends JpaRepository<Personal, Integer>{
+ 
 }
